get orders

+++ src/app/modules/user.services.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-expressions */
 import { TUser } from './user.interface'
 import User from './user.model'
 
@@ -10,17 +11,77 @@ const getAllUser = async () => {
   return result
 }
 const getSingleUser = async (id: string) => {
-  const result = await User.findById(id)
-  return result
+  if (await User.isUserExists(id)) {
+    const result = await User.findById(id)
+
+    return result
+  } else {
+    return
+  }
 }
+
 const updateUser = async (id: string, userData: TUser) => {
-  console.log(userData)
-  const result = await User.findByIdAndUpdate(id, userData, { new: true })
-  return result
+  try {
+    if (await User.isUserExists(id)) {
+      const result = await User.findByIdAndUpdate(id, userData, { new: true })
+      return result
+    } else {
+      throw new Error('missing user2')
+    }
+  } catch (error) {
+    throw new Error('missing user1')
+  }
 }
+
 const deleteUser = async (id: string) => {
-  const result = await User.deleteOne({ _id: id })
-  return result
+  try {
+    if (await User.isUserExists(id)) {
+      const result = await User.deleteOne({ _id: id })
+      return result
+    }
+  } catch (error) {
+    throw new Error('missing is absent')
+  }
+}
+
+const AddNewProduct = async (id: string, userData: TUser) => {
+  const addproduct = await User.updateOne(
+    { _id: id },
+    {
+      $push: { orders: userData },
+    },
+    { new: true },
+  )
+  return addproduct
+}
+
+const getOrder = async (id: string) => {
+  try {
+    if (await User.isUserExists(id)) {
+      const result = await User.findById(id)
+      return result
+    }
+  } catch (error) {
+    throw new Error('invalid user')
+  }
+}
+
+const totalPrice = async (id: string) => {
+  try {
+    if (await User.isUserExists(id)) {
+      const userData = await User.findById(id).select('orders')
+      const orders = userData?.orders
+      let sum = 0
+      // eslint-disable-next-line @typescript-eslint/no-unused-vars, no-unused-vars
+      const totalPrice = orders?.map((order) => {
+        const total = order.price * order.quantity
+        sum = sum + total
+      })
+      return sum
+    }
+  } catch (error) {
+    throw new Error('no total price')
+  }
 }
 
 export const userServices = {
@@ -29,4 +90,7 @@ export const userServices = {
   getSingleUser,
   deleteUser,
   updateUser,
+  AddNewProduct,
+  getOrder,
+  totalPrice,
 }
